Revert "[add] BaseClass Track.h & namespace"
